# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  totalStudents: number;
  totalCourses: number;
  activeSemesters: number;
  attendanceAtRisk: number;
  pendingFees: number;
  assignmentsDue: number;
  syntheticLabel: string;
}

export interface StudentSummary {
  studentId: string;
  name: string;
  email: string;
  department: string;
  year: number;
  avatarSeed: string;
}

export interface StudentProfile {
  studentId: string;
  name: string;
  email: string;
  phone: string;
  department: string;
  year: number;
  dateOfBirth: string;
  enrollmentDate: string;
  syntheticLabel: string;
}

export interface SemesterReport {
  semester: string;
  academicYear: string;
  gpa: number;
  credits: number;
  status: string;
}

export interface AttendanceReport {
  courseCode: string;
  courseName: string;
  attended: number;
  held: number;
  percentage: number;
  shortage: number;
  status: string;
}

export interface FeeReport {
  semester: string;
  total: number;
  paid: number;
  pending: number;
  status: string;
}

export interface AssignmentReport {
  title: string;
  courseCode: string;
  dueDate: string;
  /** @nullable */
  score: number | null;
  maxMarks: number;
  status: string;
}

export interface ExamReport {
  title: string;
  courseCode: string;
  examDate: string;
  room: string;
}

export interface ResultReport {
  courseCode: string;
  courseName: string;
  internalMarks: number;
  examMarks: number;
  totalMarks: number;
  grade: string;
  gradePoint: number;
}

export type StudentReportTotals = {
  feeTotal: number;
  feePaid: number;
  feePending: number;
  attendancePercentage: number;
  attendanceShortage: number;
};

export interface StudentReport {
  profile: StudentProfile;
  semesters: SemesterReport[];
  attendance: AttendanceReport[];
  fees: FeeReport[];
  assignments: AssignmentReport[];
  upcomingExams: ExamReport[];
  results: ResultReport[];
  totals: StudentReportTotals;
  syntheticLabel: string;
}

export interface AssistantQueryInput {
  /** @minLength 1 */
  message: string;
  /** @nullable */
  studentId?: string | null;
}

export interface AssistantQueryResponse {
  answer: string;
  intent: string;
  report: StudentReport | null;
  matches: StudentSummary[];
  needsClarification: boolean;
  syntheticLabel: string;
}

export type ListStudentsParams = {
search?: string;
studentId?: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};