# lib/api-zod/src/generated/types/assistantQueryInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AssistantQueryInput {
  /** @minLength 1 */
  message: string;
  /** @nullable */
  studentId?: string | null;
}