# lib/api-zod/src/generated/types/assistantQueryResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudentReport } from './studentReport';
import type { StudentSummary } from './studentSummary';

export interface AssistantQueryResponse {
  answer: string;
  intent: string;
  report: StudentReport | null;
  matches: StudentSummary[];
  needsClarification: boolean;
  syntheticLabel: string;
}