# lib/api-zod/src/generated/types/attendanceReport.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttendanceReport {
  courseCode: string;
  courseName: string;
  attended: number;
  held: number;
  percentage: number;
  shortage: number;
  status: string;
}