# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalStudents: number;
  totalCourses: number;
  activeSemesters: number;
  attendanceAtRisk: number;
  pendingFees: number;
  assignmentsDue: number;
  syntheticLabel: string;
}