# lib/api-zod/src/generated/types/examReport.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExamReport {
  title: string;
  courseCode: string;
  examDate: string;
  room: string;
}