# lib/api-zod/src/generated/types/resultReport.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResultReport {
  courseCode: string;
  courseName: string;
  internalMarks: number;
  examMarks: number;
  totalMarks: number;
  grade: string;
  gradePoint: number;
}