# lib/api-zod/src/generated/types/semesterReport.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SemesterReport {
  semester: string;
  academicYear: string;
  gpa: number;
  credits: number;
  status: string;
}