# lib/api-zod/src/generated/types/studentProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentProfile {
  studentId: string;
  name: string;
  email: string;
  phone: string;
  department: string;
  year: number;
  dateOfBirth: string;
  enrollmentDate: string;
  syntheticLabel: string;
}