# lib/api-zod/src/generated/types/studentReport.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssignmentReport } from './assignmentReport';
import type { AttendanceReport } from './attendanceReport';
import type { ExamReport } from './examReport';
import type { FeeReport } from './feeReport';
import type { ResultReport } from './resultReport';
import type { SemesterReport } from './semesterReport';
import type { StudentProfile } from './studentProfile';
import type { StudentReportTotals } from './studentReportTotals';

export interface StudentReport {
  profile: StudentProfile;
  semesters: SemesterReport[];
  attendance: AttendanceReport[];
  fees: FeeReport[];
  assignments: AssignmentReport[];
  upcomingExams: ExamReport[];
  results: ResultReport[];
  totals: StudentReportTotals;
  syntheticLabel: string;
}