# lib/api-zod/src/generated/types/studentSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentSummary {
  studentId: string;
  name: string;
  email: string;
  department: string;
  year: number;
  avatarSeed: string;
}